requirements.txt
# =============================================================================
# NUST Bank LLM Assistant - distributed hybrid RAG runtime + evaluation
# =============================================================================

# --- API / frontend runtime ---
fastapi>=0.115.0
uvicorn>=0.30.0
gradio>=4.44.0
requests>=2.32.0
python-multipart>=0.0.9
python-dotenv>=1.0.0

# --- Retrieval / vector infrastructure ---
qdrant-client>=1.11.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
langchain-text-splitters>=0.3.0
numpy>=1.26.0

# --- Queue / async processing ---
celery>=5.4.0
redis>=5.0.0
pypdf>=5.0.0

# --- LLM + transformers ---
huggingface-hub>=0.24.0
transformers>=4.44.0

# --- Data prep / notebooks ---
pandas>=2.0.0
openpyxl>=3.1.0
tqdm>=4.66.0
ipywidgets>=8.0.0
jupyter>=1.0.0
ipykernel>=6.29.0

# --- Backend orchestration / security ---
langgraph>=0.2.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0

# --- Tests ---
pytest>=8.0.0
